Finished game functionality and styling
Wow. What a doozy that was! I have finished all functionality for
playing the game, and styled the page. I have also added functionality
for resetting the game after completing it (this is what has taken me
so long between commits). I also added some try catches to help catch
some errors that i noticed while testing the game. It is hopefully
fool proof at this point. All that is left to do is add some extra
credit functionality for making the computer a little more challenging
to play against!